Create Quantrra.py
Please note that this code is for educational purposes only and should not be used for any malicious activities. The code assumes that the necessary libraries are installed and accessible to the Python interpreter.


The code uses the socket module to get the IP address, the platform module to get information about the device, and the subprocess module to execute the v4l2-ctl command to get information about the camera.

This code captures GPS location and camera information using the appropriate libraries and modules. It also includes the necessary import statements and combines the functions from both snippets.